[REFACTOR] VerificationCodeRepository 패키지 변경 (#102)
