+++ src/main/java/ject/componote/domain/auth/dao/impl/InMemoryVerificationCodeRepository.java
@@ -1,12 +1,14 @@
-package ject.componote.infra.mail.repository.impl;
+package ject.componote.domain.auth.dao.impl;
 
-import ject.componote.infra.mail.model.VerificationCode;
-import ject.componote.infra.mail.repository.VerificationCodeRepository;
+import ject.componote.domain.auth.dao.VerificationCodeRepository;
+import ject.componote.domain.auth.domain.VerificationCode;
+import org.springframework.stereotype.Repository;
 
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
+@Repository
 public class InMemoryVerificationCodeRepository implements VerificationCodeRepository {
     private final Map<String, VerificationCode> storage = new ConcurrentHashMap<>();
